+++ src/app/public/ipc.js
@@ -1,27 +1,50 @@
 // vendors
 // project
-const { createLogger } = require("@podman-desktop-companion/logger");
-const {
-  isApiRunning,
-  resetSystem,
-  startApi,
-  getSystemEnvironment,
-  getSystemConnections,
-  getSystemInfo,
-  pruneSystem,
-  getMachines,
-  createApiRequest,
-  connectToContainer,
-  createMachine,
-  connectToMachine,
-  restartMachine,
-  stopMachine,
-  removeMachine,
-  getUserConfiguration,
-  setUserConfiguration
-} = require("@podman-desktop-companion/container-client");
+const { getLevel, setLevel, createLogger } = require("@podman-desktop-companion/logger");
+const userSettings = require("@podman-desktop-companion/user-settings");
+const { Backend } = require("@podman-desktop-companion/container-client");
 // locals
 const logger = createLogger("shell.ipc");
+const backend = new Backend();
+
+async function getProgram() {
+  const name = await backend.getProgramName();
+  return {
+    name,
+    path: await backend.getProgramPath(),
+    currentVersion: await backend.getProgramVersion(),
+    title: name,
+    homepage: `https://${name}.io`
+  }
+}
+
+async function getUserConfiguration() {
+  const options = {
+    engine: await backend.getEngine(),
+    program: await getProgram(),
+    autoStartApi: userSettings.get("autoStartApi", false),
+    minimizeToSystemTray: userSettings.get("minimizeToSystemTray", false),
+    communication: "api",
+    path: userSettings.getPath(),
+    logging: {
+      level: getLevel()
+    }
+  };
+  return options;
+}
+
+async function setUserConfiguration(options) {
+  Object.keys(options).forEach(async (key) => {
+    if (key === "logging.level") {
+      setLevel(options[key]);
+    } else if (key === "engine") {
+      await backend.setEngine(options[key]);
+    } else {
+      userSettings.set(key, options[key]);
+    }
+  });
+  return await getUserConfiguration();
+}
 
 const servicesMap = {
   "/container/engine/request": async function (options) {
@@ -32,7 +55,7 @@ const servicesMap = {
       statusText: "API request error"
     };
     try {
-      const response = await createApiRequest(options);
+      const response = await backend.createApiRequest(options);
       result = {
         ok: response.status >= 200 && response.status < 300,
         data: response.data,
@@ -41,7 +64,7 @@ const servicesMap = {
         statusText: response.statusText
       };
     } catch (error) {
-      logger.error("API request error", error);
+      logger.error("API request error", error.message);
       result.statusText = `API request error: ${error.message}`;
     }
     return result;
@@ -54,46 +77,52 @@ const servicesMap = {
     return await getUserConfiguration();
   },
   "/system/running": async function () {
-    return await isApiRunning();
+    return await backend.getIsApiRunning();
   },
   "/system/connections": async function () {
-    return await getSystemConnections();
+    return await backend.getSystemConnections();
   },
   "/system/info": async function () {
-    return await getSystemInfo();
+    return await backend.getSystemInfo();
   },
   "/system/prune": async function () {
-    return await pruneSystem();
+    return await backend.pruneSystem();
   },
   "/system/environment": async function () {
-    return await getSystemEnvironment();
+    const program = await getProgram();
+    const system = await backend.getSystemEnvironment();
+    system.userConfiguration = await getUserConfiguration();
+    const hasProgram = program.path && program.currentVersion;
+    const isRunning = await backend.getIsApiRunning();
+    system.provisioned = hasProgram && isRunning;
+    return system;
   },
   "/system/api/start": async function () {
-    return await startApi();
+    return await backend.startApi();
   },
   "/system/reset": async function () {
-    return await resetSystem();
+    return await backend.resetSystem();
   },
   "/container/connect": async function ({ Id }) {
-    return await connectToContainer(Id);
+    return await backend.connectToContainer(Id);
   },
   "/machines/list": async function () {
-    return await getMachines();
+    return await backend.getMachines();
   },
   "/machine/restart": async function ({ Name }) {
-    return await restartMachine(Name);
+    return await backend.restartMachine(Name);
   },
   "/machine/stop": async function ({ Name }) {
-    return await stopMachine(Name);
+    return await backend.stopMachine(Name);
   },
   "/machine/connect": async function ({ Name }) {
-    return await connectToMachine(Name);
+    return await backend.connectToMachine(Name);
   },
   "/machine/remove": async function ({ Name, force }) {
-    return await removeMachine(Name, force);
+    return await backend.removeMachine(Name, force);
   },
   "/machine/create": async function (opts) {
-    return await createMachine(opts);
+    return await backend.createMachine(opts);
   }
 };
 
@@ -102,7 +131,7 @@ module.exports = {
     invoke: async (method, params) => {
       let result = {
         success: false,
-        data: null,
+        data: undefined,
         warnings: []
       };
       const service = servicesMap[method];